Get value from key first; if not existing, try recursive get
Due to splitting object field by dot (.), when an user provides a data
which key has a dot, it will cause an issue for an `undefined` data
which looks like not rendered.

So, to solve this, we should try to get the object value from the key
first. If it's not defined or null, then we will try to do the original
recursive way to get value.